backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
groq==0.11.0
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
